perf: reuse VM stack across match_at calls and use u32 for Unicode ranges
Two optimizations that dramatically reduce execution overhead:

1. Reuse match stack and capture-group arrays via MatchArg instead of
   heap-allocating 3 Vecs per match_at call. C uses alloca + reuse for
   the same purpose (regexec.c STACK_INIT). This eliminates 3 malloc +
   3 free per match attempt.

2. Store multi-byte character class ranges as Vec<u32> instead of Vec<u8>,
   converting once at compile time. The binary search in is_in_code_range
   now uses direct u32 indexing (matching C's OnigCodePoint* arithmetic)
   instead of per-iteration byte deserialization.

Results (before → after vs C):
- Lookbehind: 2.65-2.79x → 1.01-1.07x (at parity with C)
- Unicode \p{Greek}: 2.96x → 1.17x
- Unicode \p{Cyrillic}: 3.67x → 1.19x
- Named captures: 3.59x → 1.65x
- Case-insensitive phrase: 1.06x → 0.87x (now Rust wins)
- 10-branch alternation: 1.34x → 1.06x

+++ src/regint.rs
@@ -373,10 +373,10 @@ pub enum OperationPayload {
         bsp: Box<BitSet>,
     },
     CClassMb {
-        mb: Vec<u8>,
+        mb: Vec<u32>,
     },
     CClassMix {
-        mb: Vec<u8>,
+        mb: Vec<u32>,
         bsp: Box<BitSet>,
     },
     AnyCharStarPeekNext {
